Move RemindAtAlarmManagerImpl.kt to Hilt
