+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/domain/IRemindAtAlarmManager.kt
@@ -1,16 +1,12 @@
 package com.realityexpander.tasky.agenda_feature.domain
 
-import android.content.Context
-
 interface IRemindAtAlarmManager {
 
     fun setAlarmsForAgendaItems(
-        context: Context,
         agendaItems: List<AgendaItem>
     )
 
     fun cancelAllAlarms(
-        context: Context,
         onFinished: () -> Unit = {},
     )
 }
