+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/presentation/agenda_screen/AgendaScreenEvent.kt
@@ -65,6 +65,6 @@ sealed interface AgendaScreenEvent {
         data class NavigateToOpenReminder(val reminderId: TaskId) : AgendaScreenEvent, OneTimeEvent
         data class NavigateToEditReminder(val reminderId: TaskId) : AgendaScreenEvent, OneTimeEvent
 
-        data class SetAllAgendaItemAlarms(val agendaItems: List<AgendaItem>) : AgendaScreenEvent, OneTimeEvent
+//        data class SetAllAgendaItemAlarms(val agendaItems: List<AgendaItem>) : AgendaScreenEvent, OneTimeEvent
     }
 }
