+++ app/src/main/java/com/realityexpander/tasky/core/presentation/broadcastReceivers/CompleteTaskBroadcastReceiver.kt
@@ -5,8 +5,8 @@ import android.content.Context
 import android.content.Intent
 import androidx.core.app.NotificationManagerCompat
 import com.realityexpander.tasky.agenda_feature.data.repositories.agendaRepository.agendaRepositoryImpls.AgendaRepositoryImpl
-import com.realityexpander.tasky.core.presentation.notifications.RemindAtNotificationManager
-import com.realityexpander.tasky.core.presentation.notifications.RemindAtNotificationManager.ALARM_NOTIFICATION_INTENT_ACTION_COMPLETE_TASK
+import com.realityexpander.tasky.core.presentation.notifications.RemindAtNotificationManagerImpl
+import com.realityexpander.tasky.core.presentation.notifications.RemindAtNotificationManagerImpl.ALARM_NOTIFICATION_INTENT_ACTION_COMPLETE_TASK
 import com.realityexpander.tasky.core.util.UuidStr
 import com.realityexpander.tasky.core.util.toIntegerHashCodeOfUUIDString
 import dagger.hilt.android.AndroidEntryPoint
@@ -24,7 +24,7 @@ class CompleteTaskBroadcastReceiver : BroadcastReceiver() {
         intent ?: return
         if(intent.action != ALARM_NOTIFICATION_INTENT_ACTION_COMPLETE_TASK) return
 
-        val agendaItemId = intent.getStringExtra(RemindAtNotificationManager.ALARM_NOTIFICATION_INTENT_EXTRA_AGENDA_ID) as? UuidStr
+        val agendaItemId = intent.getStringExtra(RemindAtNotificationManagerImpl.ALARM_NOTIFICATION_INTENT_EXTRA_AGENDA_ID) as? UuidStr
         agendaItemId ?: return
 
         // If AgendaItem.Task, Set `isDone` to true
